feat: Remove old Output component & Rename NewOutput to Output
