+++ client/src/app/components/stardust/Output.tsx
@@ -1,34 +1,27 @@
-/* eslint-disable max-len */
-import { BASIC_OUTPUT_TYPE, UnitsHelper } from "@iota/iota.js-stardust";
-import { WriteStream } from "@iota/util.js-stardust";
+import { BASIC_OUTPUT_TYPE, ALIAS_OUTPUT_TYPE, FOUNDRY_OUTPUT_TYPE, NFT_OUTPUT_TYPE,
+    TREASURY_OUTPUT_TYPE, SIMPLE_TOKEN_SCHEME_TYPE } from "@iota/iota.js-stardust";
+import classNames from "classnames";
 import React, { Component, ReactNode } from "react";
-import { Link } from "react-router-dom";
-import { ClipboardHelper } from "../../../helpers/clipboardHelper";
-import { NameHelper } from "../../../helpers/stardust/nameHelper";
-import MessageButton from "../MessageButton";
-import { OutputState } from "../OutputState";
+import { ReactComponent as DropdownIcon } from "./../../../assets/dropdown-arrow.svg";
+import FeatureBlock from "./FeatureBlock";
 import { OutputProps } from "./OutputProps";
+import { OutputState } from "./OutputState";
+import UnlockCondition from "./UnlockCondition";
 
 /**
  * Component which will display an output.
  */
 class Output extends Component<OutputProps, OutputState> {
     /**
-     * Create a new instance of Output.
+     * Create a new instance of NewOutput.
      * @param props The props.
      */
     constructor(props: OutputProps) {
         super(props);
 
-        const GENESIS_HASH = "0".repeat(64);
-
-        const writeStream = new WriteStream();
-        writeStream.writeUInt16("transactionId", props.output.outputIndex);
-
         this.state = {
-            formatFull: false,
-            isGenesis: props.output.messageId === GENESIS_HASH,
-            outputHash: props.output.transactionId + writeStream.finalHex()
+            output: props.output,
+            showOutputDetails: -1
         };
     }
 
@@ -38,111 +31,160 @@ class Output extends Component<OutputProps, OutputState> {
      */
     public render(): ReactNode {
         return (
-            <div className="output">
-                <div className="card--header">
-                    <h2>{this.props.advancedMode ? NameHelper.getOutputTypeName(this.props.output.output.type) : "Output"} {this.props.index}</h2>
-                </div>
-                <div className="card--content">
-                    <div className="card--label">
-                        Message Id
-                    </div>
-                    <div className="card--value row middle">
-                        {this.state.isGenesis && (
-                            <span>Genesis</span>
-                        )}
-                        {!this.state.isGenesis && (
+            <div className="output margin-l-t">
+                {this.state.output.type === ALIAS_OUTPUT_TYPE && (
+                    <React.Fragment>
+                        <div className="card--label">
+                            Alias id:
+                        </div>
+                        <div className="card--value row">
+                            {this.state.output.aliasId}
+                        </div>
+                        <div className="card--label">
+                            State index:
+                        </div>
+                        <div className="card--value row">
+                            {this.state.output.stateIndex}
+                        </div>
+                        <div className="card--label">
+                            State metadata:
+                        </div>
+                        <div className="card--value row">
+                            {this.state.output.stateMetadata}
+                        </div>
+                        <div className="card--label">
+                            Foundry counter:
+                        </div>
+                        <div className="card--value row">
+                            {this.state.output.foundryCounter}
+                        </div>
+                    </React.Fragment>
+                )}
+
+                {this.state.output.type === NFT_OUTPUT_TYPE && (
+                    <React.Fragment>
+                        <div className="card--label">
+                            Nft id:
+                        </div>
+                        <div className="card--value row">
+                            {this.state.output.nftId}
+                        </div>
+                    </React.Fragment>
+                )}
+
+                {this.state.output.type === FOUNDRY_OUTPUT_TYPE && (
+                    <React.Fragment>
+                        <div className="card--label">
+                            Serial number:
+                        </div>
+                        <div className="card--value row">
+                            {this.state.output.serialNumber}
+                        </div>
+                        <div className="card--label">
+                            Token tag:
+                        </div>
+                        <div className="card--value row">
+                            {this.state.output.tokenTag}
+                        </div>
+                        <div className="card--label">
+                            Token scheme type:
+                        </div>
+                        <div className="card--value row">
+                            {this.state.output.tokenScheme.type}
+                        </div>
+                        {this.state.output.tokenScheme.type === SIMPLE_TOKEN_SCHEME_TYPE && (
                             <React.Fragment>
-                                <Link
-                                    to={
-                                        `/${this.props.network
-                                        }/message/${this.props.output.messageId}`
-                                    }
-                                    className="margin-r-t"
-                                >
-                                    {this.props.output.messageId}
-                                </Link>
-                                <MessageButton
-                                    onClick={() => ClipboardHelper.copy(
-                                        this.props.output.messageId
-                                    )}
-                                    buttonType="copy"
-                                    labelPosition="top"
-                                />
+                                <div className="card--label">
+                                    Minted tokens:
+                                </div>
+                                <div className="card--value row">
+                                    {this.state.output.tokenScheme.mintedTokens}
+                                </div>
+                                <div className="card--label">
+                                    Melted tokens:
+                                </div>
+                                <div className="card--value row">
+                                    {this.state.output.tokenScheme.meltedTokens}
+                                </div>
+                                <div className="card--label">
+                                    Maximum supply:
+                                </div>
+                                <div className="card--value row">
+                                    {this.state.output.tokenScheme.maximumSupply}
+                                </div>
                             </React.Fragment>
                         )}
-                    </div>
-                    {this.props.advancedMode && (
-                        <React.Fragment>
-                            <div className="card--label">
-                                Transaction Id
-                            </div>
-                            <div className="card--value row middle">
-                                {this.state.isGenesis && (
-                                    <span>Genesis</span>
-                                )}
-                                {!this.state.isGenesis && (
-                                    <React.Fragment>
-                                        <span className="margin-r-t">
-                                            <Link
-                                                to={
-                                                    `/${this.props.network
-                                                }/search/${this.state.outputHash}`
-                                                }
-                                                className="margin-r-t"
-                                            >
-                                                {this.props.output.transactionId}
-                                            </Link>
-                                        </span>
-                                        <MessageButton
-                                            onClick={() => ClipboardHelper.copy(
-                                                this.props.output.transactionId
-                                            )}
-                                            buttonType="copy"
-                                            labelPosition="top"
-                                        />
-                                    </React.Fragment>
-                                )}
-                            </div>
-                            <div className="card--label">
-                                Index
-                            </div>
-                            <div className="card--value">
-                                {this.props.output.outputIndex}
-                            </div>
-                            <div className="card--label">
-                                Is Spent
-                            </div>
-                            <div className="card--value">
-                                {this.props.output.isSpent ? "Yes" : "No"}
-                            </div>
-                        </React.Fragment>
-                    )}
-                    {(this.props.output.output.type === BASIC_OUTPUT_TYPE) && (
-                        <React.Fragment>
-                            <div className="card--label">
-                                Amount
-                            </div>
-                            <div className="card--value">
-                                <button
-                                    type="button"
-                                    onClick={() => this.setState(
-                                        {
-                                            formatFull: !this.state.formatFull
-                                        }
-                                    )}
-                                >
-                                    {this.state.formatFull
-                                        ? `${this.props.output.output.amount} i`
-                                        : UnitsHelper.formatBest(Number(this.props.output.output.amount))}
-                                </button>
-                            </div>
-                        </React.Fragment>
+                    </React.Fragment>
+                )}
+
+                {/* all output types except Treasury have common output conditions */}
+                {this.state.output.type !== TREASURY_OUTPUT_TYPE && (
+                    <React.Fragment>
+                        {this.state.output.unlockConditions.map((unlockCondition, idx) => (
+                            <UnlockCondition
+                                key={idx}
+                                unlockCondition={unlockCondition}
+                            />
+                        ))}
+                        {this.state.output.featureBlocks.map((featureBlock, idx) => (
+                            <FeatureBlock
+                                key={idx}
+                                featureBlock={featureBlock}
+                            />
+                        ))}
+                        {this.state.output.type !== BASIC_OUTPUT_TYPE && this.state.output.immutableFeatureBlocks && (
+                            <React.Fragment>
+                                {this.state.output.immutableFeatureBlocks.map((immutableFeatureBlock, idx) => (
+                                    <FeatureBlock
+                                        key={idx}
+                                        featureBlock={immutableFeatureBlock}
+                                    />
+                                ))}
+                            </React.Fragment>
                         )}
-                </div>
+                        {this.state.output.nativeTokens.map((token, idx) => (
+                            <React.Fragment key={idx}>
+                                <div className="native-token padding-t-s">
+                                    <div
+                                        className="card--content__input card--value row middle"
+                                        onClick={() => this.setState({ showOutputDetails:
+                                            this.state.showOutputDetails === 1 ? -1 : 1 })}
+                                    >
+                                        <div className={classNames("margin-r-t", "card--content__input--dropdown",
+                                            "card--content__flex_between",
+                                            { opened: this.state.showOutputDetails === 1 })}
+                                        >
+                                            <DropdownIcon />
+                                        </div>
+                                        <div className="card--label">
+                                            Native token
+                                        </div>
+                                    </div>
+                                    {this.state.showOutputDetails === 1 && (
+                                    <div className="margin-l-t">
+                                        <div className="card--label">
+                                            Token id:
+                                        </div>
+                                        <div className="card--value row">
+                                            {token.id}
+                                        </div>
+                                        <div className="card--label">
+                                            Amount:
+                                        </div>
+                                        <div className="card--value row">
+                                            {token.amount}
+                                        </div>
+                                    </div>
+                                    )}
+                                </div>
+                            </React.Fragment>
+                        ))}
+                    </React.Fragment>
+                )}
+
             </div>
         );
     }
 }
 
 export default Output;
-
