+++ client/src/app/components/stardust/OutputProps.tsx
@@ -1,34 +1,28 @@
-import { IOutputResponse } from "@iota/iota.js-stardust";
-import * as H from "history";
+import { OutputTypes } from "@iota/iota.js-stardust";
 
 export interface OutputProps {
     /**
-     * The network to lookup.
-     */
-    network: string;
-
-    /**
      * The output id.
      */
-    id: string;
+    id?: string;
 
     /**
-     * The output id.
+     * The index within the parent.
      */
     index: number;
 
     /**
      * The output to display.
      */
-    output: IOutputResponse;
+    output: OutputTypes;
 
     /**
-     * History for navigation.
+     * The amount to display.
      */
-    history: H.History;
+    amount: number;
 
     /**
-     * Display advanced mode.
+     * Should hide the label.
      */
-     advancedMode: boolean;
+    hideLabel?: boolean;
 }
